requirements.txt
discord.py
python-dotenv
flask
pytz